Rename api/src/schema/schema.ts to ../schema/index.ts
