+++ api/src/index.ts
@@ -2,7 +2,7 @@ import dotenv from 'dotenv';
 
 import { create_app } from './create_app.ts';
 import { get_env } from './env.ts';
-import { schema } from './schmea/schema.ts';
+import { schema } from './schmea/index.ts';
 
 dotenv.config({ path: '.env' }); // relative to the call point, e.g. the service root
